Improve the resiliency of OpenPGPBatchJob
The OpenPGP BatchJob will continue processing even if it encounters an error with a single file in the source folder. In the event of a file processing error, the job will record the error in the log and proceed with processing the remaining files and subfolders in the source folder.