Implemented a fix to rectify Keys Not found errors if the email address of the OpenPGP Key in the KeyRing is not fully in lower-cap letters.
Implemented a fix to rectify "Cannot find Sender's or Recipient's PGP Key in the keyring" errors if the email address of the OpenPGP Key in the KeyRing is not fully in lower-cap letters.

Added the Sender and Recipient Email Addresses that are being searched for in the "Cannot find Sender's or Recipient's PGP Key in the keyring" error logs to aid in troubleshooting. Note that the email address are cap-sensitive!